Se realiza depuración del codigo, se estructra el MVC del proyecto
