+++ huevos_la_granja/lib/view/pages/administrador/adminAjuste.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-import 'witgets/drawer.dart';
-import 'witgets/drawer_administrador.dart';
+import '../../widgets/drawer_administrador.dart';
 
 //import '../widgets/drawer.dart';
 
-class AdminInventarioPage extends StatelessWidget {
-  const AdminInventarioPage({super.key});
+class AdminAjustePage extends StatelessWidget {
+  const AdminAjustePage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Administrador -Inventario"),
+        title: const Text("Administrador -Ajuste"),
       ),
       drawer: const DrawerWidgetAdministrador(email: "", name: ""),
     );
