+++ huevos_la_granja/lib/view/pages/administrador/adminInventario.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-import 'witgets/drawer.dart';
-import 'witgets/drawer_administrador.dart';
+import '../../widgets/drawer_administrador.dart';
 
 //import '../widgets/drawer.dart';
 
-class AdminVentasPage extends StatelessWidget {
-  const AdminVentasPage({super.key});
+class AdminInventarioPage extends StatelessWidget {
+  const AdminInventarioPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Administrador -Ventas"),
+        title: const Text("Administrador -Inventario"),
       ),
       drawer: const DrawerWidgetAdministrador(email: "", name: ""),
     );
