+++ huevos_la_granja/lib/view/pages/administrador/adminVentas.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-import 'witgets/drawer.dart';
-import 'witgets/drawer_administrador.dart';
+import '../../widgets/drawer_administrador.dart';
 
 //import '../widgets/drawer.dart';
 
-class AdminAjustePage extends StatelessWidget {
-  const AdminAjustePage({super.key});
+class AdminVentasPage extends StatelessWidget {
+  const AdminVentasPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Administrador -Ajuste"),
+        title: const Text("Administrador -Ventas"),
       ),
       drawer: const DrawerWidgetAdministrador(email: "", name: ""),
     );
