+++ huevos_la_granja/lib/view/pages/auxiliar_bodega/auxConsulta.dart
@@ -1,18 +1,17 @@
 import 'package:flutter/material.dart';
 
-import 'witgets/drawer.dart';
-import 'witgets/drawer_auxbodega.dart';
+import '../../widgets/drawer_auxbodega.dart';
 
 //import '../widgets/drawer.dart';
 
-class AuxIngresoPage extends StatelessWidget {
-  const AuxIngresoPage({super.key});
+class AuxConsultaPage extends StatelessWidget {
+  const AuxConsultaPage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Auxiliar - Ingreso"),
+        title: const Text("Auxiliar - Consulta"),
       ),
       drawer: const DrawerWidgetAuxbodega(email: "", name: ""),
     );
